It is in progress. More Bear.
Bear chasing is now more accurate, now i am doing code in an attempt to make custom patrol paths so the bear dont need to find a random point in the navmesh to go to. Doing code with increment path index. Dropping this for now so that I can work with level building with artists